requirements.txt
streamlit
pandas
futures
requests
mysql-connector
importlib
